Require at least one character in unquoted text
Previously, an empty string was valid for both unquoted-text and
quoted-text. While this was fine for quoted-text (it's surrounded by
quotes), this didn't work work for unquoted-text as this would make an
empty string a valid unquoted-pattern (and so also a valid pattern and
for a valid value). Finally this makes

identifier =

a valid message. This disagrees with the current FTL tinker
implementation.

Quoted-text has been updated to retain the symmetry with unquoted-text.
Quoted-pattern has been updated so that

identifier = ""

is still a valid message.