Change Annotation to { code, args, message, span }
Annotations now have unique codes: E0000 for errors, W0000 for warnings
and H0000 for hints.  Args is an array of strings with additional arguments
used for building the message of the Annotation.

Annotations also now have a span, similar to Entries.  If an annotation is only
related to a single position in the file, the Span's start and end field should
be equal.