Not backporting support for tags, documenting via test.
Tags were used to signal grammatical properties in 0.4 and
earlier. The move to attributes on terms was both backwards
and forwards incompatible, so we can't write a parser that
supports the old and current behavior.
Documenting that by creating a test that exemplifies that
the syntax that used to be a tag is now parsed as part of
a Value or Attribute value.

This is OK as we never practically used tags in projects.