Don't include the trailing line break in Messages, Terms and Comments
Revert earlier changes which added the trailing newline to the
definition of the Message, Term and Comment, as well as to the content
of Comments. The trailing line break should not be part of the
production, span nor the content.

Make Junk subclass SyntaxNode directly, rather than Entry. Entry is now
an abstract class for useful AST nodes carrying content: Messages, Terms
and Comments.

Remove the annotations field from Entry togather with the add_annotation
method and move it directly onto Junk.