Add more cases to validation.
This is adding more options to more combinators.
There's one thing I didn't find out how to validate, and that
repeat1 should cover 1 and 2, and I don't think that's possible
to validate. I tried tracing the production calls in a different
patch, but that didn't return anything useful as it doesn't count
if the 2-match ends up in a parser error somewhere else in the
message.